# Optimized fracSecsToISOString and added writer overload