@JackStouffer 2018-02-14
This diff is a little cleaner https://github.com/dlang/phobos/pull/6176/files?w=1